refactor: core/node/bitswap.go (#8349)
- extracts bitswap init to separate file and defines implicit defaults as consts
- fixes a typo where BitswapTaskWorkerCount was 128 (instead of 8)
  and BitswapEngineBlockstoreWorkerCount was 8 (instead of 128)
